// decEncCommon.d.ts
export declare const enum EncodedType {
    Number0 = 0,
    Number63 = 63,
    String0 = 64,
    String63 = 127,
    Map0 = 128,
    Map31 = 159,
    Array0 = 160,
    Array31 = 191,
    PositiveInt1 = 192,
    PositiveInt4 = 195,
    NegativeInt1 = 196,
    NegativeInt4 = 199,
    String1 = 200,
    String4 = 203,
    Map1 = 204,
    Map4 = 207,
    Array1 = 208,
    Array4 = 211,
    Double = 212,
    False = 213,
    True = 214,
    Null = 215,
    EndOfBlock = 216,
    NumberM32 = 224,
    NumberM1 = 255,
}